Move TextInput Flow types to their own file
